Layout_TestCase/http/tests/xmlhttprequest/access-control-preflight-async-not-supported-expected.txt

CONSOLE MESSAGE: XMLHttpRequest cannot load http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests/xmlhttprequest/resources/access-control-preflight-denied-xsrf.php. Origin http://http://172.16.7.123/test/jetpacktest/Layout_TestCase/http/tests is not allowed by Access-Control-Allow-Origin.
PASS: Request successfully blocked.
